move CustomTestType and CustomExtension out of tests init
