+++ asdf/tests/test_schema.py
@@ -9,7 +9,8 @@ from numpy.testing import assert_array_equal
 import asdf
 from asdf import config_context, constants, extension, get_config, resolver, schema, tagged, types, util, yamlutil
 from asdf.exceptions import AsdfConversionWarning, AsdfDeprecationWarning, AsdfWarning
-from asdf.tests import CustomExtension, helpers
+from asdf.tests import helpers
+from asdf.tests.objects import CustomExtension
 
 
 class TagReferenceType(types.CustomType):
